+++ ChangeLog.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- `e3nn.normalize_function` now uses a deterministic (not pseudorandom) algorithm to compute the normalization factor.
 
 ## [0.17.2] - 2023-03-26
 ### Fixed