+++ ChangeLog.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `LinearSHTP` module implementing the optimized linear mixing of inputs tensor product with spherical harmonics
 
 ## [0.17.3] - 2023-04-06
 ### Changed